requirements.txt
scikit-learn 
spacy 
pdfminer.six 
streamlit
flask
werkzeug

